refactor(config): rename personal_config to user_config
