+++ src/mcp_coder/utils/__init__.py
@@ -25,7 +25,7 @@ from .git_operations import (
     stage_specific_files,
 )
 from .log_utils import log_function_call, setup_logging
-from .personal_config import get_config_file_path, get_config_value
+from .user_config import get_config_file_path, get_config_value
 from .subprocess_runner import (
     CommandOptions,
     CommandResult,
@@ -53,7 +53,7 @@ __all__ = [
     # Logging utilities
     "log_function_call",
     "setup_logging",
-    # Personal configuration
+    # User configuration
     "get_config_file_path",
     "get_config_value",
     # Subprocess operations
